requirements.txt
numpy>=2,<3
pandas>=2,<3
requests>=2.32.3
psycopg[binary]~=3.2.10
alpaca-py>=0.11.0
pytz>=2024.1
attrs>=23.2.0
psycopg
hmmlearn>=0.3.0
scipy>=1.12.0
